Add tenantId filtering to Skills

+++ optashift-employee-rostering-gwtui/src/main/java/org/optaplanner/openshift/employeerostering/gwtui/client/roster/RosterListPanel.java
@@ -31,7 +31,7 @@ import org.optaplanner.openshift.employeerostering.shared.roster.RosterRestServi
 @Templated
 public class RosterListPanel implements IsElement {
 
-    private Long tenantId = -1L;
+    private Integer tenantId = -1;
 
     @Inject @DataField
     private Button refreshButton;
